+++ crates/db-storage/src/schema.rs
@@ -276,6 +276,7 @@ table! {
         phone -> Nullable<Varchar>,
         tenant_id -> Uuid,
         tariff_id -> Uuid,
+        tariff_status -> Tariff_status,
     }
 }